@@ -46,7 +46,7 @@
     "singleQuote": true
   },
   "devDependencies": {
-    "@babel/core": "^7.17.8",
+    "@babel/core": "^7.17.9",
     "@babel/eslint-parser": "^7.17.0",
     "@babel/plugin-transform-runtime": "^7.17.0",
     "@babel/preset-env": "^7.16.11",
@@ -54,15 +54,15 @@
     "@commitlint/cli": "^16.2.3",
     "@commitlint/config-conventional": "^16.2.1",
     "@fancyapps/ui": "^4.0.26",
-    "@types/inquirer": "^8.2.0",
+    "@types/inquirer": "^8.2.1",
     "@types/jest": "^27.4.1",
-    "@types/node": "^17.0.23",
+    "@types/node": "^17.0.24",
     "@types/semver": "^7.3.9",
-    "@typescript-eslint/eslint-plugin": "^5.16.0",
-    "@typescript-eslint/parser": "^5.16.0",
+    "@typescript-eslint/eslint-plugin": "^5.19.0",
+    "@typescript-eslint/parser": "^5.19.0",
     "@vue/eslint-config-prettier": "7.0.0",
     "@vue/eslint-config-typescript": "10.0.0",
-    "apidoc": "^0.51.0",
+    "apidoc": "^0.51.1",
     "babel-jest": "^27.5.1",
     "babel-loader": "^8.2.4",
     "commitizen": "^4.2.4",
@@ -70,25 +70,25 @@
     "cross-env": "^7.0.3",
     "css-loader": "^6.7.1",
     "cz-conventional-changelog": "^3.3.0",
-    "eslint": "^8.12.0",
+    "eslint": "^8.13.0",
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-prettier": "^4.0.0",
     "eslint-plugin-react": "^7.29.4",
-    "eslint-plugin-vue": "^8.5.0",
+    "eslint-plugin-vue": "^8.6.0",
     "file-loader": "^6.2.0",
     "html-webpack-plugin": "^5.5.0",
     "husky": "^7.0.4",
     "inquirer": "^8.2.2",
     "jest": "^27.5.1",
     "lerna": "^4.0.0",
-    "lint-staged": "^12.3.7",
-    "marked": "^4.0.12",
-    "npm-check-updates": "^12.5.4",
+    "lint-staged": "^12.3.8",
+    "marked": "^4.0.14",
+    "npm-check-updates": "^12.5.9",
     "postcss-loader": "^6.2.1",
-    "prettier": "^2.6.1",
-    "rollup": "^2.70.1",
-    "rollup-plugin-dts": "^4.2.0",
-    "sass": "^1.49.9",
+    "prettier": "^2.6.2",
+    "rollup": "^2.70.2",
+    "rollup-plugin-dts": "^4.2.1",
+    "sass": "^1.50.0",
     "sass-loader": "^12.6.0",
     "sort-package-json": "^1.55.0",
     "style-loader": "^3.3.1",
@@ -97,12 +97,12 @@
     "ts-node": "^10.7.0",
     "typescript": "^4.6.3",
     "url-loader": "^4.1.1",
-    "vercel": "^24.0.1",
+    "vercel": "^24.1.0",
     "vue-loader": "^17.0.0",
-    "webpack": "^5.70.0",
+    "webpack": "^5.72.0",
     "webpack-bundle-analyzer": "^4.5.0",
     "webpack-cli": "^4.9.2",
-    "webpack-dev-server": "^4.7.4",
+    "webpack-dev-server": "^4.8.1",
     "webpack-merge": "^5.8.0"
   },
   "apidoc": {
