@@ -1,6 +1,6 @@
 {
   "name": "@waline/admin",
-  "version": "0.12.2",
+  "version": "0.17.1",
   "description": "management system for waline",
   "keywords": [
     "waline",
@@ -46,13 +46,14 @@
     "@rematch/core": "^2.2.0",
     "@svgr/webpack": "^6.2.1",
     "classnames": "^2.3.1",
-    "i18next": "^21.6.14",
+    "i18next": "^21.6.16",
     "i18next-browser-languagedetector": "^6.1.4",
     "md5": "^2.3.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-i18next": "^11.16.2",
-    "react-redux": "^7.2.6",
-    "react-router-dom": "^6.2.2"
+    "qrcode.react": "^3.0.1",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
+    "react-i18next": "^11.16.6",
+    "react-redux": "^8.0.0",
+    "react-router-dom": "^6.3.0"
   }
 }
