@@ -1,3 +1,4 @@
+const speakeasy = require('speakeasy');
 const jwt = require('jsonwebtoken');
 const helper = require('think-helper');
 const { PasswordHash } = require('phpass');
@@ -17,7 +18,7 @@ module.exports = class extends BaseRest {
   }
 
   async postAction() {
-    const { email, password } = this.post();
+    const { email, password, code } = this.post();
     const user = await this.modelInstance.select({ email });
 
     if (think.isEmpty(user) || /^verify:/i.test(user[0].type)) {
@@ -33,13 +34,24 @@ module.exports = class extends BaseRest {
       return this.fail();
     }
 
-    let avatar = user[0].avatar;
+    const twoFactorAuthSecret = user[0]['2fa'];
+    if (twoFactorAuthSecret) {
+      const verified = speakeasy.totp.verify({
+        secret: twoFactorAuthSecret,
+        encoding: 'base32',
+        token: code,
+        window: 2,
+      });
+      if (!verified) {
+        return this.fail();
+      }
+    }
 
+    let avatar = user[0].avatar;
     if (/(github)/i.test(avatar)) {
       avatar =
         this.config('avatarProxy') + '?url=' + encodeURIComponent(avatar);
     }
-
     user[0].avatar = avatar;
 
     return this.success({
