@@ -13,7 +13,10 @@ module.exports = function () {
     window.SITE_URL = ${JSON.stringify(process.env.SITE_URL)};
     window.SITE_NAME = ${JSON.stringify(process.env.SITE_NAME)};
     </script>
-    <script src="https://cdn.jsdelivr.net/npm/@waline/admin"></script>
+    <script src="${
+      process.env.WALINE_ADMIN_MODULE_ASSET_URL ||
+      'https://cdn.jsdelivr.net/npm/@waline/admin'
+    }"></script>
   </body>
 </html>`;
   };
